src/requirements.txt
PyYAML<7.0,>=6.0.3
python-json-logger<5.0,>=4.0.0
prometheus-client<1.0,>=0.24.1
grpcio<2.0,>=1.78.0
grpcio-tools<2.0,>=1.78.0
grpcio-reflection<2.0,>=1.78.0
grpcio-health-checking<2.0,>=1.78.0
protobuf<7.0,>=6.32.1